Moved PEvent and PEventHandler to new file PEvent.ts
