+++ src/shared/models/PEventEmitter.ts
@@ -2,7 +2,7 @@
  * Interfaces, classes, and functions related to events and event emitters
  */
 
-import { PEvent, PEventHandler } from '@shared/util/PapiUtil';
+import { PEvent, PEventHandler } from '@shared/models/PEvent';
 
 /**
  * Event manager - accepts subscriptions to an event and runs the subscription callbacks when the event is emitted
