+++ src/shared/util/PapiUtil.ts
@@ -126,16 +126,6 @@ export const createSafeRegisterFn = <TParam extends Array<unknown>, TReturn>(
 
 // #endregion
 
-/** Callback function that accepts an event and should run when an event is emitted */
-export type PEventHandler<T> = (event: T) => void;
-
-/**
- * Function that subscribes the provided callback to run when this event is emitted.
- * @param callback function to run with the event when it is emitted
- * @returns unsubscriber function to run to stop calling the passed-in function when the event is emitted
- */
-export type PEvent<T> = (callback: PEventHandler<T>) => Unsubscriber;
-
 /**
  * Type of object passed to a complex request handler that provides information about the request.
  * This type is used as the public-facing interface for requests
